1-sprint-mission: 프로젝트 구조 재정리, ServiceFactory 클래스에서 싱글톤 패턴 제거, 출력문 정리
